Feature/for feed

머지 풀 리퀘스트 드립니다.

+++ src/main/java/com/kh/eco/comment/model/vo/CommentVO.java
@@ -10,7 +10,7 @@ import lombok.Value;
 public class CommentVO {
 
 	private Long commentNo;
-	private Long refMno;
+	private Integer refMno;
 	private Long refBno;
 	private String commentContent;
 	private Date regDate;
